Creating a Poll app using PHP
Creating a poll app using PHP.  App is not that close to being done.
Most functionality not completed yet.  User votes are being added to a
database but the user can not see that anything is happening.  However,
as of now if they do refresh the page they will never see the same two
teams twice.  Keeping the database size smaller for now because it
makes testing easier